Reimplement the tracker client part using QSparql which can use direct-access mode
The QSparql or Qt5Sparql direct-access mode QTRACKER_DIRECT gets a direct connect
with the sqlite3 library through libtracker-sparql. This is a lot more efficient
than using the DBusMessage based Resources D-Bus object of Tracker (which is
actually only to be used for debugging with for example D-Feet and/or dbus-send).

This is only done for the SparqlQuery part as the GraphUpdated part can only happen
over the Resources D-Bus object of Tracker (and is in itself optimized).

See QSparql's examples for more information and see the Steroids1 and direct-access
mode features of libtracker-sparql and tracker-store for more information.